Order

+++ src/main/java/pl/sda/gda15/project/domain/entity/User.java
@@ -26,7 +26,7 @@ public class User {
     private String city;
     private String country;
     private String email;
-    private Integer tel;
+    private Long tel;
 
 
 }
